PsychoJL.TextStim — Type

TextStim()

Constructor for a TextStim object

Constructor inputs:

win::Window

textMessage::String.........default = "future text"

pos::Vector{Int64}........position: default = [10,10]

Optional constructor inputs:

color::Vector{Int64}.........default = (128, 128, 128)

fontName::String = "",

fontSize::Int64 = 12,.........default = 12

scale::Float64 = 1.0,.........not the same as font size

font::Any.........default is taken from Winow

horizAlignment::Int64.........*default = 1, 0 = center, -1 = right

vertAlignment::Int64 = 1.........*default = 1, 0 = center, -1 = bottom

style::String.........default = "normal", options include "bold" and "italic"

orientation.........orientation in degrees

Methods:

draw()

Notes: Using different font sizes requires loading them as different fonts. For now it is easier to load a large version of a font and using scale to scale the size of the resulting image.

draw(various shape types) - Method

Draws the shape (Line, Rect, Ellipse, TextStim, etc.) into the back buffer.

Example:

	newRect = Rect(myWin, 
			100,			# width
			100, 			# height
			[200,200],		# position
			lineColor = [255,0,0], 
			fillColor = [255,128,128] 
			)
	draw(newRect) 		# in PsychoPy this would have been newRect.draw()

"""