PsychoJL.Window — Type

Window()

Constructor for a Window object

Constructor inputs:

size::MVector{2, Int64}

fullScreen::Bool

Optional constructor inputs:

color::MVector{3, Int64}

colorSpace::String .......Future. Not implemented yet

pos::MVector{2, Float64} ......position

timeScale::String .......default = "milliseconds"

title::String ......default = "Window"

Full list of fields

win::Ptr{SDL_Window}

size::MVector{2, Int64}

pos::MVector{2, Int64} ......position

color::MVector{3, Int64}

colorSpace::String

renderer::Ptr{SDL_Renderer}

font::Ptr{SimpleDirectMediaLayer.LibSDL2.TTFFont}

boldFont::Ptr{SimpleDirectMediaLayer.LibSDL2.TTFFont}

italicFont::Ptr{SimpleDirectMediaLayer.LibSDL2.TTFFont}

event::Base.RefValue{SDL_Event}

fullScreen::Bool

timeScale::String .......defaults is milliseconds. Other option is seconds

title::String

startTime::Float64 .......global proximy for startTime() and stopTime()

Methods:

closeAndQuitPsychoJL()

closeWinOnly()

flip()

getPos()

getSize()

hideWindow()

mouseVisible()

setFullScreen()

PsychoJL.closeAndQuitPsychoJL — Method

closeAndQuitPsychoJL(win::Window)

Attempts to close a PsychoJL Window and quit SDL.

PsychoJL.closeWinOnly — Method

closeWinOnly(win::Window)

Attempts to close a PsychoJL Window without quiting SDL.

PsychoJL.flip — Method

flip(win::Window)

Flips the offscreen buffer on screen. In other words, all of the visual objects that you have drawn offscreen prior to the flip will now become visible.

PsychoJL.getPos — Method

getPos(win::Window)

Returns the center of the window. This, as well as the dimensions, can chage when going to full screen

PsychoJL.getSize — Method

getSize(win::Window)

Returns the width and height of the window. Dimensions can chage when going to full screen.

PsychoJL.hideWindow — Method

hideWindow(win::Window)

Attempts to hide a PsychoJL Window.

PsychoJL.mouseVisible — Method

mouseVisible(mode::Bool)

Hides or shows the cursor

PsychoJL.setFullScreen — Method

setFullScreen(win::Window, mode::Bool)

Allows you to flip between windowed and full-screen mode.