PsychExpAPIs.TextStim — Type

TextStim()

Constructor for a TextStim object

Constructor inputs:

win::Window

textMessage::String.........default = "future text"

pos::Vector{Int64}........position: default = [10,10]

Optional constructor inputs:

color::PsychoColor.........default = (128, 128, 128)

fontName::String = "",

fontSize::Int64 = 12,.........default = 12

scale::Float64 = 1.0,.........not the same as font size

font::Any.........default is taken from Window

horizAlignment::Int64.........default = -1, 0 = center, +1 = right

vertAlignment::Int64 = 1.........default = 1, 0 = center, -1 = bottom

style::String.........default = "normal", options include "bold" and "italic"

orientation.........orientation in degrees

Methods:

draw()

Notes: Using different font sizes requires loading them as different fonts. For now it is easier to load a large version of a font and using scale to scale the size of the resulting image.

PsychExpAPIs.draw — Method

draw(text::TextStim; wrapLength::Int64)

Draws an TextStim to the back buffer.

Inputs:

text::TextStim

Optional Inputs:

wrapLength::Int64......in pixels (for now)